requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
PyMuPDF
pytesseract
pillow